Modularize addStrict.
