+++ tfjs-core/src/ops/add_strict.ts
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright 2020 Google Inc. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+import {Tensor} from '../tensor';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+import * as util from '../util';
+
+import {add} from './add';
+import {op} from './operation';
+
+/**
+ * Adds two `tf.Tensor`s element-wise, A + B.
+ *
+ * Inputs must be the same shape. For broadcasting support, use add() instead.
+ *
+ * @param a The first Tensor to add element-wise.
+ * @param b The second Tensor to add element-wise.
+ */
+function addStrict_<T extends Tensor>(a: T|TensorLike, b: T|TensorLike): T {
+  const $a = convertToTensor(a, 'a', 'addStrict');
+  const $b = convertToTensor(b, 'b', 'addStrict');
+  util.assertShapesMatch($a.shape, $b.shape, 'Error in addStrict: ');
+  return add($a, $b);
+}
+
+export const addStrict = op({addStrict_});
