+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -24,8 +24,8 @@ import {ALL_ENVS, describeWithFlags} from '../../jasmine_util';
 // flexibility to change in future.
 
 const CHAINED_OPS = [
-  'add', 'batchNorm', 'broadcastTo', 'div', 'divNoNan', 'oneHot', 'pad',
-  'square', 'tile', 'transpose'
+  'add', 'addStrict', 'batchNorm', 'broadcastTo', 'div', 'divNoNan', 'oneHot',
+  'pad', 'square', 'tile', 'transpose'
 ];
 
 describeWithFlags('chained ops', ALL_ENVS, () => {
